+++ src/Exception.php
@@ -18,6 +18,7 @@ class Exception extends \Exception
      */
     private $params = [];
 
+    /** @var array */
     public $trace2; // because PHP's use of final() sucks!
 
     /**
@@ -42,6 +43,26 @@ class Exception extends \Exception
         $this->trace2 = debug_backtrace(DEBUG_BACKTRACE_PROVIDE_OBJECT);
     }
 
+    /**
+     * Change message (subject) of a current exception. Primary use is
+     * for localization purposes.
+     *
+     * @param string $message
+     *
+     * @return $this
+     */
+    public function setMessage($message)
+    {
+        $this->message = $message;
+
+        return $this;
+    }
+
+    /**
+     * Return trace array.
+     *
+     * @return array
+     */
     public function getMyTrace()
     {
         return $this->trace2;